Refator interface code to improve the readability
