+++ src/interface.rs
@@ -1,1 +1,2 @@
 pub mod injector;
+mod verifier;
